Add support for custom metadata
Custom metadata can now be added to options by storing an object on the
`metadata` key of options with the key:value pairs of metadata.

The keys will automatically have 'x-awz-meta-' added to them, so they
should just have the name wanted for the metadata without the aws
specific prefix.